Refactor/column

+++ src/main/java/kg/peaksoft/taskTrackerb6/db/service/CardService.java
@@ -33,6 +33,7 @@ public class CardService {
     private final CardConverter converter;
     private final NotificationRepository notificationRepository;
     private final BasketRepository basketRepository;
+    private final WorkspaceRepository workspaceRepository;
 
     private User getAuthenticateUser() {
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
@@ -107,7 +108,7 @@ public class CardService {
         if (card.getIsArchive().equals(true)) {
             if (baskets != null) {
                 for (Basket b : baskets) {
-                    if (b.getCard().equals(card)) {
+                    if (b.getCard() != null && b.getCard().equals(card)) {
                         basketRepository.deleteBasket(b.getId());
                     }
                 }
@@ -180,6 +181,7 @@ public class CardService {
         card.setColumn(column);
         column.addCard(card);
         card.setCreatedAt(LocalDate.now());
+        user.addCard(card);
         Card save = cardRepository.save(card);
         return converter.convertToCardInnerPageResponse(save);
     }
@@ -194,51 +196,39 @@ public class CardService {
                 }
         );
 
-        if (!card.getCreator().equals(user)) {
-            throw new BadCredentialException("You can not archive this card!");
-        }
+        Column column = columnRepository.findById(card.getColumn().getId()).orElseThrow(
+                () -> new NotFoundException("Column with id: " + id + " not found!")
+        );
 
-        card.setIsArchive(!card.getIsArchive());
-        if (card.getIsArchive().equals(true)) {
-            Basket basket = new Basket();
-            basket.setCard(card);
-            basket.setArchivedUser(user);
-            basketRepository.save(basket);
-        }
+        Workspace workspace = workspaceRepository.findById(column.getBoard().getWorkspace().getId()).orElseThrow(
+                () -> new NotFoundException("Workspace with id: " + column.getBoard().getWorkspace().getId() + " not found!")
+        );
 
-        if (card.getIsArchive().equals(false)) {
-            List<Basket> baskets = user.getBaskets();
-            if (baskets != null) {
-                for (Basket b : baskets) {
-                    if (b.getCard() != null && b.getCard().equals(card)) {
-                        basketRepository.deleteBasket(b.getId());
-                    }
+        if (card.getCreator().equals(user) || column.getCreator().equals(user) || workspace.getLead().equals(user)) {
+
+            card.setIsArchive(!card.getIsArchive());
+            if (card.getIsArchive().equals(true)) {
+                Basket basket = new Basket();
+                basket.setCard(card);
+                basket.setArchivedUser(user);
+                card.setBasket(basket);
+                basketRepository.save(basket);
+            }
+
+            if (card.getIsArchive().equals(false)) {
+                List<Basket> baskets = basketRepository.findAll();
+                    for (Basket b : baskets) {
+                        if (b.getCard() != null && b.getCard().equals(card)) {
+                            basketRepository.deleteBasket(b.getId());
+                        }
                 }
             }
+        } else {
+            System.out.println("In card service!");
+            throw new BadCredentialException("You can not archive this card!");
         }
 
         log.info("Card with id: {} successfully archived", id);
         return converter.convertToCardInnerPageResponse(card);
-    }
-
-
-    public ArchiveResponse getAllArchivedCardsByBoardId(Long id) {
-        Board board = boardRepository.findById(id).orElseThrow(
-                () -> new NotFoundException("Board with id: " + id + " not found!")
-        );
-
-        ArchiveResponse archiveResponse = new ArchiveResponse();
-        List<Column> columns = board.getColumns();
-        List<CardResponse> archivedCardResponse = new ArrayList<>();
-        for (Column column : columns) {
-            for (Card card : column.getCards()) {
-                if (card.getIsArchive().equals(true)) {
-                    archivedCardResponse.add(converter.convertToResponseForGetAll(card));
-                }
-            }
-        }
-
-        archiveResponse.setCardResponses(archivedCardResponse);
-        return archiveResponse;
     }
 }
